optiload-ai/backend/simulation-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.6
kombu==5.3.4
numpy==1.26.3
scipy==1.12.0
httpx==0.26.0
prometheus-client==0.19.0
python-dotenv==1.0.0
